Handbag Cloud API/phamchri_project/requirements.txt
Flask==2.0.1
google-cloud-datastore==2.1.6
json2html==1.3.0
python-jose
flask-cors
python-dotenv
requests
authlib